requirements.txt
mysql-connector-python
ttkthemes
python-dotenv
messagebox
validate-docbr